Separated target success and target search success reducers
